refactor: page router 수정
